adding modified calculator prgm

+++ IdeaProjects/java_august/src/september16/CalculatorPrgm.java
@@ -3,8 +3,8 @@ package september16;
 public class CalculatorPrgm {
 
     public static void main(String[] args) {
-        int num1 = 10;
-        int num2 = 5;
+        double num1 = 10.0;
+        double num2 = 5.0;
         char operator = '*';
 
         double result;
